requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.1